+++ src/main/java/de/wwu/scdh/teilsp/services/extensions/ConfigurablePlugin.java
@@ -3,6 +3,7 @@ package de.wwu.scdh.teilsp.services.extensions;
 import java.util.Map;
 
 import de.wwu.scdh.teilsp.exceptions.ConfigurationException;
+import de.wwu.scdh.teilsp.config.ExtensionConfiguration;
 
 
 /**
@@ -35,4 +36,27 @@ public interface ConfigurablePlugin {
      * @return an array of argument descriptors
      */
     public ArgumentDescriptor<?>[] getArgumentDescriptor();
+
+    /**
+     * A descriptor for the special argument that is always present
+     * and stores the node type.
+     *
+     * See {@link de.wwu.scdh.teilsp.config.ExtensionConfiguration#NODE_TYPES}
+     * for node types.
+     */
+    public static final ArgumentDescriptor<String> SPECIAL_ARGUMENT_NODE_TYPE =
+	new ArgumentDescriptorImpl<String>
+	(String.class, "__oxbytei_nodeType__",
+	 "The special argument that stores the node type.",
+	 ExtensionConfiguration.NODE_TYPES);
+
+    /**
+     * A descriptor for the special argument that is always present
+     * and stores the node name.
+     */
+    public static final ArgumentDescriptor<String> SPECIAL_ARGUMENT_NODE_NAME =
+	new ArgumentDescriptorImpl<String>
+	(String.class, "__oxbytei_nodeName__",
+	 "The special argument that stores the node name.");
+
 }